search: fully load assumptions
When solving with assumptions, test whether all of them can be assigned in
on iteration without calling propagate before. In case the formula under
assumptions is satisfiable, this results in a faster solution.

If the assumptions together do not work, for now jump back to the default
behavior. Instead, the conflict information could also be used already.
This is not implemented here, and left for future investigation.