limits: add simplification step limit
For some formula, full simplification takes too long compared to the
expected benefit. Therefore, a step limit is introduced, which allows to
interrupt simplification in a deterministic way.

By default, the limit is deactivated. However, with the introduced command
line option "max-simp-steps", the limit can be set, influencing the
simplification time.