drat: fix deleting bogus clauses
Currently, the solver only prints DRUP proofs, where deleting clauses
that do not occur is not relevant. Hence, the version until now
produced proofs that contained an unsatisfiable core. However, proof
checker reported spurious clause deletions.

This commit fixes the issue, which is caused by the solver shrinking
clauses in case of falsified literals. This change was not covered
in the proof.

Furthermore, we add an assertion to test whether we are using the
proof API correctly, i.e. only drop literals in case we add a clause
to the proof - which is used when strengthening a clause.