var_decay: allow to dynamically increase value
In a Glucose version, it has been noticed that gradually moving
the var_decay from a smaller to a larger value is good for
performance. This way, both values that are strong for unsatisfiable
and satisfiable formulas are available during search.

The constants to move the value have been taken from Glucose. Every
5000 conflicts, the var_decay value is increased, by 0.01.