chrono: enable non-chronological backtracking
During backtracking, decide whether non-chronological backtracking
should be used, or default backtracking. This is decided based on
the number of conflicts that already happened, as well as the
backjump distance.

This technique allows to speed up search in case many large jumps
happen, as propagation results are kept on the trail, at the cost
of having the literals on the trail being unsorted.